+++ libs/designer/src/lib/ui/configuretemplate/resources.ts
@@ -134,6 +134,16 @@ export const useResourceStrings = () => {
       id: 'JBa1qe',
       description: 'The label for the workflow display name',
     }),
+    LastModified: intl.formatMessage({
+      defaultMessage: 'Last modified',
+      id: 'lx0teD',
+      description: 'Last modified label',
+    }),
+    CreatedDate: intl.formatMessage({
+      defaultMessage: 'Created date',
+      id: 'iuh4i3',
+      description: 'Created date label',
+    }),
     State: intl.formatMessage({
       defaultMessage: 'State',
       id: 'IG4XXf',
@@ -214,15 +224,33 @@ export const useResourceStrings = () => {
       id: 'an1Z0T',
       description: 'Label for the workflow images section',
     }),
+    WorkflowImagesDescription: intl.formatMessage({
+      defaultMessage:
+        'Provide your workflow images in both Azure light and dark themes. Open the workflow in the designer, take screenshots in each theme, and save the images. Then upload them to Azure Blob Storage and share the shared access signature (SAS) link.',
+      id: 'B2TJcx',
+      description: 'Description for the workflow images section',
+    }),
     LightModeImage: intl.formatMessage({
       defaultMessage: 'Light mode image',
       id: '5LLxEp',
       description: 'Light mode image label',
     }),
+    LightModeImageDescription: intl.formatMessage({
+      defaultMessage:
+        'Provide your workflow image in the Azure light theme. Upload the image to Azure Blob Storage and share the shared access signature (SAS) link.',
+      id: 'Za33CQ',
+      description: 'Light mode image description',
+    }),
     DarkModeImage: intl.formatMessage({
       defaultMessage: 'Dark mode image',
       id: 'CdxaGY',
       description: 'Dark mode image label',
+    }),
+    DarkModeImageDescription: intl.formatMessage({
+      defaultMessage:
+        'Provide your workflow image in the Azure dark theme. Upload the image to Azure Blob Storage and share the shared access signature (SAS) link.',
+      id: 'NziQUu',
+      description: 'Dark mode image description',
     }),
     InAppKind: intl.formatMessage({
       defaultMessage: 'In-app',