+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/tabs/parametersTab/index.tsx
@@ -88,8 +88,9 @@ import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import { ConnectionInline } from './connectionInline';
 import { ConnectionsSubMenu } from './connectionsSubMenu';
-import { isAgentConnector } from '../../../../../common/utilities/Utils';
 import { useCognitiveServiceAccountDeploymentsForNode } from '../../../connectionsPanel/createConnection/custom/useCognitiveService';
+import { isAgentConnectorAndAgentServiceModel, isAgentConnectorAndDeploymentId } from './helpers';
+import { useShouldEnableFoundryServiceConnection } from './hooks';
 
 // TODO: Add a readonly per settings section/group
 export interface ParametersTabProps extends PanelTabProps {
@@ -239,6 +240,7 @@ const ParameterSection = ({
   const isTrigger = useSelector((state: RootState) => isRootNodeInGraph(nodeId, 'root', state.workflow.nodesMetadata));
   const operationInfo = useOperationInfo(nodeId);
   const dependencies = useDependencies(nodeId);
+  const isFoundryServiceConnectionEnabled = useShouldEnableFoundryServiceConnection();
 
   // Specific for agentic scenarios
   const { data: deploymentsForCognitiveServiceAccount } = useCognitiveServiceAccountDeploymentsForNode(nodeId, operationInfo?.connectorId);
@@ -268,6 +270,10 @@ const ParameterSection = ({
     () => rootState.operations.inputParameters[nodeId] ?? { parameterGroups: {} },
     [nodeId, rootState.operations.inputParameters]
   );
+
+  const isAgentServiceConnection = useMemo(() => {
+    return isAgentConnectorAndAgentServiceModel(operationInfo.connectorId, group.id, nodeInputs.parameterGroups);
+  }, [group.id, nodeInputs.parameterGroups, operationInfo.connectorId]);
 
   const onValueChange = useCallback(
     (id: string, newState: ChangeState, skipStateSave?: boolean) => {
@@ -673,6 +679,9 @@ const ParameterSection = ({
 
   const settings: Settings[] = group?.parameters
     .filter((x) => !x.hideInUI && shouldUseParameterInGroup(x, group.parameters))
+    .filter((param) => {
+      return param.parameterName !== 'agentModelType' || isFoundryServiceConnectionEnabled;
+    })
     .map((param) => {
       const { id, label, value, required, showTokens, placeholder, editorViewModel, dynamicData, conditionalVisibility, validationErrors } =
         param;
@@ -721,6 +730,7 @@ const ParameterSection = ({
           pickerCallbacks: getPickerCallbacks(param),
           tokenpickerButtonProps: {
             location: panelLocation === PanelLocation.Left ? TokenPickerButtonLocation.Right : TokenPickerButtonLocation.Left,
+            hideButtonOptions: { hideDynamicContent: isAgentServiceConnection, hideExpression: isAgentServiceConnection },
           },
           agentParameterButtonProps: { showAgentParameterButton },
           hostOptions: {
@@ -781,10 +791,6 @@ const getSubMenu = (parameter: ParameterInfo) => {
     return <ConnectionsSubMenu />;
   }
   return null;
-};
-
-const isAgentConnectorAndDeploymentId = (id: string, key: string): boolean => {
-  return isAgentConnector(id) && equals(key, 'inputs.$.deploymentId', true);
 };
 
 export const getEditorAndOptions = (